Added explicity operator= and copy constructor when other one already present
In gcc-9 the warning -Wdeprecated-copy [1] is now enabled by -Wextra, and this
becomes an error when using -Werror as in CMakeLists.txt. This means that the
"implicit declaration of a copy constructor or copy assignment operator" is now
an error.

This commit adds copy constructors and assignment operators for those objects
where the other one is already explicitly declared. Copy constructors use the
parent copy constructor, assignment operators call the parent assignment operator.

In Kassiopeia/Trajectories/*/KSTraj*Particle classes the pointers to field functions
are copied in the initializer list instead of regenerated, e.g.:
```
fGetMagneticFieldPtr( aParticle.fGetMagneticFieldPtr )
```
Similarly the KSMathArray parent fields are copied for those classes.

[1] https://gcc.gnu.org/onlinedocs/gcc-9.1.0/gcc/C_002b_002b-Dialect-Options.html#index-Wdeprecated-copy