Convert warning to info message when adding additional controls/terms to trajectory
The message was introduced to warn the user in the special case when terms/controls of the same type are added twice (e.g. when a term_propagation is added to a trajectory with a term_propagation). However, with the current implementation the user is warned even when a term/control of different type is added to the trajectory.

This change converts the warning message to an info message, which only states that an additional term/control is added to the trajectory.