kasperenv.sh: Fix PATH pollution
Previously, code that was intended to remove old KASPER entries from PATH variables was broken, this is fixed. The code didn't work since it was introduced in January 2020 due to using single quotes and therefore avoiding variable replacement. Since "/" is also a letter that is frequently used in paths, to avoid future problems the letter was replaced with ";", which `thisroot.sh` also uses for this cause.